added helper functions to formatting folder
