--- src/helpers/rounding.jsx
@@ -1,7 +1,0 @@
-  // helper function to round ratings to nearest tenths place
-
-export function roundTenths(num, place) {
-    if( !place) place = 0;
-    let pow = Math.pow(10,place);
-    return Math.round(num*pow)/pow;
-}
